refactor: Wrap `ProofCollection` into variant of new enum `SingleProof`
