+++ src/models/blockchain/transaction/validity/mod.rs
@@ -3,5 +3,6 @@ pub mod collect_type_scripts;
 pub mod kernel_to_outputs;
 pub mod proof_collection;
 pub mod removal_records_integrity;
+pub mod single_proof;
 pub mod tasm;
 pub mod transaction_validity;
